Change dart->c++ interface to ask for chunks of rows.
In Dart RawReceivePort's are not garbage collected. This means we
have to be careful about properly disposing of them.

In the previous design the user could open a stream and then never use it simple by writing:

stream = db.getItems().

This caused a RawReceive port to be opened. But this port is never garbage collected and so
neither is the stream returned by this call.

Instead move to a design where dart asks for chunks of rows to be returned. We create a new
port for every chunk and every chunk is guaranteed to return (possibly with an error). This means
all RawReceivePorts will be closed and all referenced objects will be garbage collected.

With the streaming design there was no way