Move to synchronous api for performance
Whilst it is theoretically nice to have all db operations async leveldb is so
fast that the overhead of the async opereraions (creating dart port, futures,
etc...) substantially reduces the db performance.

Therefore change the majority of the api to be synchronous.

For now opening and closing the db is still async.