Correct race conditions when closing a DB
It is possible that unreferenceDB is called before db->thread is initialized
if a 2nd thread quickly takes a reference to a shared db and then drops it.
However the initializing thread still has a reference so it is safe to call
pthread_join() if the refcount was 0

The actual closing of the db and its file descriptors must be run whilst
the shared lock is taken so that any threads attempting to open the same file will
succeed.